+++ releases.md
@@ -1,5 +1,11 @@
 # Releases
 
+## Unreleased
+
+  - Add `Capture#each_object_id` for getting object identifiers.
+  - Add `retained_addresses:` option to `Sampler#analyze` to capture addresses.
+  - Add `Sampler#analyze(retained_minimum: 100)` - if the retained_size is less than this, the analyse won't proceed.
+
 ## v1.4.0
 
   - Implement [Cooper-Harvey-Kennedy](https://www.cs.tufts.edu/~nr/cs257/archive/keith-cooper/dom14.pdf) algorithm for finding root objects in memory leaks.